+++ plugins/markdown/plugin-shiki/src/node/markdown/applyHighlighter/index.ts
@@ -19,7 +19,7 @@ export const applyHighlighter = async (
   {
     langs = bundledLanguageNames,
     langAlias = {},
-    defaultLang = 'plain',
+    defaultLang,
     transformers: userTransformers = [],
     ...options
   }: ShikiHighlightOptions = {},
@@ -61,7 +61,7 @@ export const applyHighlighter = async (
         },
         transformers: [
           ...transformers,
-          ...(options.highlightLines ?? true
+          ...((options.highlightLines ?? true)
             ? [transformerCompactLineOptions(attrsToLines(attrs))]
             : []),
           ...whitespaceTransformer(attrs, options.whitespace),