+++ themes/theme-default/src/client/layouts/NotFound.vue
@@ -13,7 +13,7 @@ const homeText = themeLocale.value.backToHome ?? 'Back to home'
 </script>
 
 <template>
-  <div class="theme-container">
+  <div class="vp-theme-container">
     <main class="page">
       <div class="theme-default-content">
         <h1>404</h1>
@@ -25,3 +25,17 @@ const homeText = themeLocale.value.backToHome ?? 'Back to home'
     </main>
   </div>
 </template>
+
+<style scoped lang="scss">
+.vp-theme-container {
+  max-width: 740px;
+  margin: 0 auto;
+  padding: 2rem 2.5rem;
+}
+
+@media (max-width: 959px) {
+  .vp-theme-container {
+    padding: 2rem;
+  }
+}
+</style>