+++ tools/helper/src/node/utils/getInstalledStatus.ts
@@ -8,7 +8,7 @@ export const getInstalledStatus = (
     pkg && createRequire(currentUrl).resolve(pkg)
 
     return true
-  } catch (error) {
+  } catch {
     return false
   }
 }